requirements.txt
streamlit
groq
gspread
google-auth
pandas